Add shared containers

+++ inc/widgets/widgets.ctrl.php
@@ -177,18 +177,6 @@ switch( $display_mode )
 		init_colorpicker_js();
 }
 
-// Get Skin used by current Blog:
-$blog_skin_ID = $Blog->get( $skin_type.'_skin_ID' );
-$SkinCache = & get_SkinCache();
-$Skin = & $SkinCache->get_by_ID( $blog_skin_ID );
-// Make sure containers are loaded for that skin:
-$skins_container_list = $Blog->get_main_containers();
-// Get widget containers from database
-$WidgetContainerCache = & get_WidgetContainerCache();
-$WidgetContainerCache->load_where( 'wico_coll_ID = '.$Blog->ID.' AND wico_skin_type = '.$DB->quote( $skin_type ) );
-$blog_container_list = $WidgetContainerCache->get_ID_array();
-
-
 /**
  * Perform action:
  */
@@ -221,7 +209,7 @@ switch( $action )
 		$Session->assert_received_crumb( 'widget' );
 
 		$WidgetContainer = & get_widget_container( $Blog->ID, $container );
-		if( !in_array( $WidgetContainer->get( 'code' ), array_keys( $skins_container_list ) ) )
+		if( ! in_array( $WidgetContainer->get( 'code' ), array_keys( $Blog->get_main_containers() ) ) )
 		{ // The container is not part of the current skin
 			$Messages->add( T_('WARNING: you are adding to a container that does not seem to be part of the current skin.'), 'error' );
 		}
@@ -543,52 +531,64 @@ switch( $action )
 		}
 		break;
 
- 	case 'list':
+	case 'list':
 		break;
 
- 	case 're-order' : // js request
- 		// Check that this action request is not a CSRF hacked request:
+	case 're-order' : // js request
+		// Check that this action request is not a CSRF hacked request:
 		$Session->assert_received_crumb( 'widget' );
 
- 		$DB->begin();
+		$DB->begin();
 
- 		$blog_container_ids = implode( ',', $blog_container_list );
- 		// Reset the current orders to avoid duplicate entry errors
-		$DB->query( 'UPDATE T_widget__widget
-			SET wi_order = wi_order * -1
-			WHERE wi_wico_ID IN ( '.$blog_container_ids.' )' );
+		// Get IDs of all widget containers:
+		$SQL = new SQL( 'Get IDs of all widget containers of colleciton #'.$Blog->ID );
+		$SQL->SELECT( 'wico_ID' );
+		$SQL->FROM( 'T_widget__container' );
+		$SQL->WHERE( 'wico_coll_ID = '.$Blog->ID );
+		$SQL->WHERE_and( 'wico_skin_type = '.$DB->quote( $skin_type ) );
+		$blog_container_IDs = $DB->get_col( $SQL );
 
-		foreach( $containers as $container_fieldset_id => $widgets )
-		{ // loop through each container and set new order
-			$WidgetContainer = & get_widget_container( $Blog->ID, $container_fieldset_id );
-			if( ( $WidgetContainer->ID == 0 ) && ( count( $widgets ) > 0 ) )
-			{ // Widget was moved to an empty main widget container, it needs to be created
-				$WidgetContainer->dbinsert();
-			}
-			$order = 0; // reset counter for this container
-			foreach( $widgets as $widget )
-			{ // loop through each widget
-				if( $widget = preg_replace( '~[^0-9]~', '', $widget ) )
-				{ // valid widget id
-					$order++;
-					$DB->query( 'UPDATE T_widget__widget
-						SET wi_order = '.$order.',
-							wi_wico_ID = '.$WidgetContainer->ID.'
-						WHERE wi_ID = '.$widget.' AND wi_wico_ID IN ( '.$blog_container_ids.' )' );	// Doh! Don't trust the client request!!
+		if( $blog_container_IDs )
+		{
+			$blog_container_IDs = $DB->quote( $blog_container_IDs );
+
+			// Reset the current orders to avoid duplicate entry errors
+			$DB->query( 'UPDATE T_widget__widget
+				SET wi_order = wi_order * -1
+				WHERE wi_wico_ID IN ( '.$blog_container_IDs.' )' );
+
+			foreach( $containers as $container_fieldset_id => $widgets )
+			{ // loop through each container and set new order
+				$WidgetContainer = & get_widget_container( $Blog->ID, $container_fieldset_id );
+				if( ( $WidgetContainer->ID == 0 ) && ( count( $widgets ) > 0 ) )
+				{ // Widget was moved to an empty main widget container, it needs to be created
+					$WidgetContainer->dbinsert();
+				}
+				$order = 0; // reset counter for this container
+				foreach( $widgets as $widget )
+				{ // loop through each widget
+					if( $widget = preg_replace( '~[^0-9]~', '', $widget ) )
+					{ // valid widget id
+						$order++;
+						$DB->query( 'UPDATE T_widget__widget
+							SET wi_order = '.$order.',
+								wi_wico_ID = '.$WidgetContainer->ID.'
+							WHERE wi_ID = '.$widget.' AND wi_wico_ID IN ( '.$blog_container_IDs.' )' );	// Doh! Don't trust the client request!!
+					}
 				}
 			}
-		}
 
-		// Cleanup deleted widgets and empty temp containers
-		$DB->query( 'DELETE FROM T_widget__widget
-			WHERE wi_order < 1
-			AND wi_wico_ID IN ( '.$blog_container_ids.' )' ); // Doh! Don't touch other blogs!
+			// Cleanup deleted widgets and empty temp containers
+			$DB->query( 'DELETE FROM T_widget__widget
+				WHERE wi_order < 1
+				AND wi_wico_ID IN ( '.$blog_container_IDs.' )' ); // Doh! Don't touch other blogs!
+		}
 
 		$DB->commit();
 
- 		$Messages->add( T_( 'Widgets updated' ), 'success' );
- 		send_javascript_message( array( 'sendWidgetOrderCallback' => array( 'blog='.$Blog->ID ) ) ); // exits() automatically
- 		break;
+		$Messages->add( T_( 'Widgets updated' ), 'success' );
+		send_javascript_message( array( 'sendWidgetOrderCallback' => array( 'blog='.$Blog->ID ) ) ); // exits() automatically
+		break;
 
 
 	case 'reload':
